requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
google-api-core==2.15.0
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.7.1
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
httplib2==0.22.0
idna==3.6
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.6.0.66
pandas==1.5.3
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tweepy==4.12.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18